+++ demos/index-all.html
@@ -109,6 +109,10 @@
                 <span class="tag layer">Points</span>
             </li>
             <li>
+                23. <a href="index-samples.html?sample=23">Legend Name Hunter</a>. Tests the legends ability to snipe a
+                layer name from the config instead of waiting for the layer to register.
+            </li>
+            <li>
                 26.
                 <a href="index-samples.html?sample=26">Fog Hilighter</a>. A demonstration of feature highlighting with
                 the fog hilighter.
@@ -256,9 +260,6 @@
         Maps that are currently being used on real sites, demonstrated with RAMP4. Note that some not all the features
         on the real sites are currently supported.
         <ul>
-            <li>
-                <a href="index-samples.html?sample=23"> 23. Climate Action Map </a>
-            </li>
             <li>
                 <a href="index-samples.html?sample=24"> 24. Canadian Environmental Sustainability Indicators </a>
             </li>
