+++ demos/index-samples.html
@@ -38,7 +38,7 @@
                 <option value="layer-metadata">20. Layer with metadata</option>
                 <option value="section-item">21. Section Items</option>
                 <option value="info-section">22. Info Sections</option>
-                <option value="cam">23. Climate Action Map (CAM)</option>
+                <option value="legname">23. Legend Name Hunt</option>
                 <option value="cesi">24. Canadian Environmental Sustainability Indicators (CESI)</option>
                 <option value="cccs">25. Canadian Centre for Climate Services (CCCS)</option>
                 <option value="fog-hilight">26. Feature Highlighting with Fog Hilighter</option>
