--- public/index-cam.html
@@ -1,23 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport" content="width=device-width,initial-scale=1.0" />
-        <title>ramp-core</title>
-        <link rel="stylesheet" href="./ramp.css" />
-    </head>
-    <body style="margin: 0">
-        <noscript>
-            <strong
-                >We're sorry but ramp-core doesn't work properly without JavaScript enabled. Please enable it to
-                continue.</strong
-            >
-        </noscript>
-
-        <div id="IE"></div>
-        <div id="app" style="height: 100vh; max-height: -webkit-fill-available"></div>
-        <script src="./ramp.browser.iife.js"></script>
-        <script type="module" src="./starter-scripts/cam.js"></script>
-    </body>
-</html>
